0018: A
Fixed integer overflow by using `long`

+++ 0018.cpp
@@ -3,19 +3,19 @@ public:
     vector<vector<int>> fourSum(vector<int>& nums, int target) {
         int n = nums.size();
         set<vector<int>> ans;
-        unordered_map<int, set<vector<int>>> umap;
+        unordered_map<long, set<vector<int>>> umap;
         for (int b = 1; b < n - 2; b++) {
             for (int a = 0; a < b; a++) {
                 vector<int> temp;
                 temp.push_back(nums[a]);
                 temp.push_back(nums[b]);
                 sort(temp.begin(), temp.end());
-                int lhs = nums[a] + nums[b];
+                long lhs = nums[a] + nums[b];
                 umap[lhs].insert(temp);
             }
             int c = b + 1;
             for (int d = b + 2; d < n; d++) {
-                int rhs = target - nums[c] - nums[d];
+                long rhs = (long)target - nums[c] - nums[d];
                 for (auto v : umap[rhs]) {
                     v.push_back(nums[c]);
                     v.push_back(nums[d]);
